refactor: move instruction section to components
